% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSEMM-package.R
\docType{package}
\name{plotSEMM}
\alias{plotSEMM}
\alias{plotSEMM-package}
\title{Graphing Nonlinear Relations Among Latent Variables from Structural Equation Mixture Models}
\description{
Graphing Nonlinear Relations Among Latent Variables from Structural Equation Mixture Models
}
\details{
Contains a graphical user interface to generate the
diagnostic plots proposed by Bauer (2005) and Pek & Chalmers (2015) to investigate
nonlinear latent variable interactions in latent regression models.

Creates plots which accompany Bauers (2005) semiparametric method of modeling
Structural Equation Mixture Models (SEMMs) by allowing researchers to visualize
potential nonlinear relationships between a latent predictor and outcome. Additionally,
a graphical user interface (GUI) is available for interactive use and is found in the function
\code{\link{plotSEMM_GUI}}.
}
\author{
Bethany Kok and Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

