\docType{package}
\name{plotSEMM}
\alias{plotSEMM}
\alias{plotSEMM-package}
\title{Graphing nonlinear relations among latent variables from Structural Equation Mixture Models}
\description{
  Graphing nonlinear relations among latent variables from
  Structural Equation Mixture Models
}
\details{
  Contains functions \code{\link{plotSEMM_setup}},
  \code{\link{plotSEMM_contour}}, and
  \code{\link{plotSEMM_probability}}. Creates plots which
  accompany Bauers (2005) semiparametric method of modeling
  Structural Equation Mixture Models (SEMMs) by allowing
  researchers to visualize potential nonlinear
  relationships between a latent predictor and outcome.
}
\author{
  Bethany Kok and Phil Chalmers
  \email{rphilip.chalmers@gmail.com}
}
\references{
  Bauer, D.J. (2005). A semiparametric approach to modeling
  nonlinear relations among latent variables. Structural
  Equation Modeling: A Multidisciplinary Journal, 12(4),
  513-535.

  Pek, J., Sterba, S., Kok, B.E. & Bauer, D.J. (in prep).
  Visualizing nonlinear relations among latent variables:
  An online plotting utility and R package.
}
\keyword{package}

