% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotSEMM_GUI}
\alias{plotSEMM_GUI}
\title{PlotSEMM GUI}
\usage{
plotSEMM_GUI()
}
\description{
Graphical user interface with the shiny package. Supports manual input as well as importing
from precomputed Mplus files.
}
\examples{
\dontrun{
plotSEMM_GUI()
}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{GUI}
\keyword{shiny}

