% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotSEMM_probability}
\alias{plotSEMM_probability}
\title{Probability plot}
\usage{
plotSEMM_probability(SEMLIdatapks, EtaName = "Eta1", lnty = 3, lncol = 1,
  title = "", leg = TRUE, ...)
}
\arguments{
\item{SEMLIdatapks}{object returned from \code{\link{plotSEMM_setup}}}

\item{EtaName}{Label of the latent predictor.  If no value is provided, defaults to Eta1.}

\item{lnty}{Determines the line types used for the class lines.  If no value is provided,
defaults to 3.  See \code{\link{par}} for information about line type.}

\item{lncol}{Determines the line colors used for the class lines.  If no value is
provided, defaults to 1.  See \code{\link{par}} for information about line type.}

\item{title}{Titles the graph.}

\item{leg}{Logical variable.  If TRUE, a legend accompanies the graph.  If FALSE,
no legend appears.  Defaults to TRUE.}

\item{...}{addition inputs, mostly from plotSEMM_GUI()}
}
\description{
Requires \code{plotSEMM_setup} be run first. Generates a plot which expresses
the mixing probabilities for each latent class conditioned on the latent predictor.
}
\examples{
\dontrun{
# 2 class empirical example on positive emotions and heuristic processing in Pek,
# Sterba, Kok & Bauer (XXXX)
pi <- c(0.602, 0.398)

alpha1 <- c(3.529, 2.317)

alpha2 <- c(0.02, 0.336)

beta21 <- c(0.152, 0.053)

psi11 <- c(0.265, 0.265)

psi22 <- c(0.023, 0.023)


plotobj <- plotSEMM_setup(pi, alpha1, alpha2, beta21, psi11, psi22)

plotSEMM_probability(plotobj)

plotSEMM_probability(plotobj , EtaName = "Latent Predictor", lnty = 2, title = "Probability")
}
}
\author{
Bethany Kok and Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{plotSEMM_setup}}, \code{\link{plotSEMM_contour}}
}
\keyword{color}
\keyword{hplot}

